requirements.txt
streamlit
google-generativeai
python-dotenv
db-sqlite3
pandas
openpyxl
matplotlib